requirements.txt
pytorch_lightning
mlflow==2.0.0rc0
protobuf~=3.19.0 # temporary solution to avoid breaking changes made to 3rd party streamlit (necessary for mlflow) https://discuss.streamlit.io/t/typeerror-descriptors-cannot-not-be-created-directly/25639
csc-mlops
numpy
torchvision
monai
itk
tqdm
pandas
matplotlib
xnat
boto3
docker
pytest
fsspec
GitPython
colorlog
pandas
minio==7.0.4
openpyxl
torchmetrics
nibabel